Fixed tests after lead provider school report journey refactor
