+++ app/views/dashboard/_navigation.html.erb
@@ -3,7 +3,7 @@
     <%= component.nav_item(path: dashboard_path) do %>
       Overview
     <% end %>
-    <%= component.nav_item(path: start_lead_providers_report_schools_path) do %>
+    <%= component.nav_item(path: lead_providers_report_schools_start_path) do %>
       Schools
     <% end %>
   <% end %>
