+++ app/views/dashboard/show.html.erb
@@ -5,7 +5,7 @@
         <%= render partial: "navigation" %>
         <h1 class="govuk-heading-xl"><%= @lead_provider.name %></h1>
         <h2 class="govuk-heading-m">
-          <%= govuk_link_to "Confirm your schools", start_lead_providers_report_schools_path %>
+          <%= govuk_link_to "Confirm your schools", lead_providers_report_schools_start_path %>
         </h2>
         <p>Choose a delivery partner and upload schools using a CSV.</p>
       </div>
