--- spec/forms/lead_provider_delivery_partner_form_spec.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-require "rails_helper"
-
-RSpec.describe LeadProviderDeliveryPartnerForm, type: :model do
-  describe "validations" do
-    it { is_expected.to validate_presence_of(:delivery_partner_id).with_message("Choose a delivery partner") }
-  end
-end
